Major refactor using Reflux stores/actions and more files
