+++ app/js/actions/HistoryActions.js
@@ -1,0 +1,6 @@
+import Reflux from 'reflux';
+
+export default Reflux.createActions([
+  'pageBack',
+  'navigateToModal'
+]);
