--- app/js/actions/PolicyTableActions.js
@@ -1,5 +1,0 @@
-import * as Reflux from 'reflux';
-
-export default Reflux.createActions([
-  'policyPointSelected'
-]);
